store html in a file on the disk; skip page fetching when debugging
new global variable: debug
used as a flag to speed up development.

new function: fetch_html
fetches the html code from a link and saves it in a file on the disk.
skip the fetching of the page if the debug flag is on and the file is
already on the disk.

new function: slugify
make a slug from a arbitrary, potentially unicode, string. used to
generate a valid filename from a url.